requirements.txt
Django
djangorestframework
django-import-export